+++ src/search/models.py
@@ -7,9 +7,10 @@ from django.urls import reverse
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
-from core.fields import ChoiceArrayField
-from aids.models import Aid
 from aids.constants import AUDIENCES_GROUPED
+from aids.models import Aid
+from core.fields import ChoiceArrayField
+from pages.models import Page
 
 
 def logo_upload_to(instance, filename):
@@ -263,3 +264,24 @@ class SearchPageLite(SearchPage):
         proxy = True
         verbose_name = "page personnalisée"
         verbose_name_plural = "pages personnalisées"
+
+
+class MinisiteTab(Page):
+    """
+    Proxy class to make Page model available for minisites
+    as a Tab.
+    """
+    class Meta:
+        proxy = True
+        verbose_name = "onglet (toutes les PP)"
+        verbose_name_plural = "onglets (toutes les PP)"
+
+
+class MinisiteTabLite(Page):
+    """
+    Proxy class to make a lite admin for ministe Tab.
+    """
+    class Meta:
+        proxy = True
+        verbose_name = "onglet"
+        verbose_name_plural = "onglets"
